pyproject.toml
[project]
name = "mpaciu-ae-cap-1-1"
version = "0.1.0"
description = "Stock Analysis Platform - DCF Valuation and Recommendations"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "duckdb>=1.4.1",
    "pytest>=9.0.0",
    "yfinance>=0.2.66",
    "streamlit>=1.30.0",
    "beautifulsoup4>=4.12.0",
    "langgraph>=0.0.20",
    "langchain-openai>=0.0.5",
    "langchain-core>=0.1.0",
    "google-api-python-client>=2.100.0",
    "playwright>=1.40.0",
    "finnhub-python>=2.4.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "requests>=2.31.0",
    "boto3>=1.41.1",
]

[tool.pytest.ini_options]
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
    "slow: marks tests as slow running tests",
]
